Panning & massive restructuring
Only real game change is the new ability to pan when either that button
is selected or using the middle mouse button. However, the code got
massively restructured and most of it condensed into a dll and config
XML file that can theoretically be used for any platforming game. The
unitypackage containing everything non-game-specific is in the main
project folder.